fix(c/driver/sqlite,c/validation): Ensure float/double values are not truncated on bind or select (#585)
Closes #578 and modifies the validation suite so that it will catch this
type of accidental cast should it happen again.

+++ c/driver/sqlite/statement_reader.c
@@ -182,8 +182,8 @@ AdbcStatusCode AdbcSqliteBinderBindNext(struct AdbcSqliteBinder* binder, sqlite3
         }
         case NANOARROW_TYPE_FLOAT:
         case NANOARROW_TYPE_DOUBLE: {
-          int64_t value = ArrowArrayViewGetDoubleUnsafe(binder->batch.children[col],
-                                                        binder->next_row);
+          double value = ArrowArrayViewGetDoubleUnsafe(binder->batch.children[col],
+                                                       binder->next_row);
           status = sqlite3_bind_double(stmt, col + 1, value);
           break;
         }
